Move DocumentId into own file, add tests
