+++ p2panda-rs/src/document/mod.rs
@@ -257,11 +257,13 @@
 
 #[allow(clippy::module_inception)]
 mod document;
+mod document_id;
 mod document_view;
 mod error;
 
 #[allow(unused_imports)]
 use document::{build_graph, reduce};
-pub use document::{Document, DocumentBuilder, DocumentId};
+pub use document::{Document, DocumentBuilder};
+pub use document_id::DocumentId;
 pub use document_view::{DocumentView, DocumentViewId};
 pub use error::{DocumentBuilderError, DocumentError, DocumentViewError};
